Remove unnecessary base classes in dependency resolution tests
